jwt system updates

+++ Backend/RRSS/src/main/java/com/demo/rrss/rrssbackend/service/ReviewService.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.ui.Model;
 import org.springframework.web.server.ResponseStatusException;
 
 import com.demo.rrss.rrssbackend.entity.Review;
@@ -22,35 +23,35 @@ public class ReviewService {
 		return response.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Review not found"));
 	}
 
-	public void addReview(ReviewRequest request) {
+	public void addReview(ReviewRequest request, Model model) {
+		Long userId = (Long) model.getAttribute("userId"); // Not used at the moment
 		Review review = new Review();
 		review.setProductId(request.getProductId());
-		review.setUserId(request.getUserId());
+		review.setUserId(userId);
 		review.setReviewData(request.getReviewData());
 		review.setPublishDate(new java.sql.Timestamp(new java.util.Date().getTime()));
 		repository.save(review);
 	}
 
-	public void updateReview(Long reviewId, ReviewRequest request) {
+	public void updateReview(Long reviewId, ReviewRequest request, Model model) {
 		Optional<Review> existingReview = repository.findById(reviewId);
-		if (existingReview.isPresent()) {
+		Long userId = (Long) model.getAttribute("userId");
+		if (existingReview.isPresent() && existingReview.get().getUserId() == userId){
 			Review review = existingReview.get();
-			review.setProductId(request.getProductId());
-			review.setUserId(request.getUserId());
 			review.setReviewData(request.getReviewData());
 			review.setPublishDate(new java.sql.Timestamp(new java.util.Date().getTime()));
-
 			repository.save(review);
 		} else {
-			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Review not found");
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Review not found or you do not have permission to update it.");
 		}
 	}
 
-	public void deleteReview(Long reviewId) {
-		if (repository.existsById(reviewId))
+	public void deleteReview(Long reviewId, Model model) {
+		Long userId = (Long) model.getAttribute("userId");
+		if (repository.existsById(reviewId) && repository.findById(reviewId).get().getUserId() == userId)
 			repository.deleteById(reviewId);
 		else
-			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Review not found");
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Review not found or you do not have permission to delete it.");
 	}
 
 	public List<Review> getAllReviews() {
